# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/rag/upload": {
      "post": {
        "description": "Returns immediately with status PENDING. Poll GET /api/rag/documents until READY or FAILED — embedding a long PDF takes tens of seconds.",
        "operationId": "RagController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "conversationId"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "conversationId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Upload a PDF for retrieval",
        "tags": [
          "rag"
        ]
      }
    },
    "/api/rag/url": {
      "post": {
        "operationId": "RagController_url",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestUrlDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Ingest a YouTube or web URL",
        "tags": [
          "rag"
        ]
      }
    },
    "/api/rag/documents": {
      "get": {
        "operationId": "RagController_list",
        "parameters": [
          {
            "name": "conversationId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Ingestion status for one conversation",
        "tags": [
          "rag"
        ]
      }
    },
    "/api/rag/documents/{id}": {
      "delete": {
        "operationId": "RagController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Remove a document, its vectors and its stored file",
        "tags": [
          "rag"
        ]
      }
    },
...
    "/api/payments/packs": {
      "get": {
        "description": "Served from the server so prices live in one place. The billing page renders whatever this returns rather than hardcoding amounts.",
        "operationId": "PaymentsController_listPacks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreditPackDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "The credit packs on offer",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/payments/order": {
      "post": {
        "operationId": "PaymentsController_createOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create a Razorpay order for a pack",
        "tags": [
          "payments"
        ]
      }
    }
  },
  "info": {
    "title": "Kurakani API",
    "description": "Backend for the Kurakani chat app.\n\nAuth is a session cookie issued by Better Auth. Its own routes live under\n`/api/auth/*` and are handled by the Better Auth handler rather than a Nest\ncontroller, so they are not listed here — the frontend talks to them through\nthe `better-auth/react` client, which ships its own types.\n\nEvery route below is protected by a global guard: no cookie means 401,\nbefore the handler runs.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local backend (direct)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local frontend (proxied by Next.js)"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Set by Better Auth on sign-in."
      }
    },
    "schemas": {
...
      "SourceType": {
        "type": "string",
        "enum": [
          "PDF",
          "YOUTUBE",
          "WEB"
        ]
      },
      "IngestStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "PROCESSING",
          "READY",
          "FAILED"
        ]
      },
      "DocumentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "conversationId": {
            "type": "string"
          },
          "sourceType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SourceType"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IngestStatus"
              }
            ]
          },
          "filename": {
            "type": "string",
            "nullable": true
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true
          },
          "error": {
            "type": "string",
            "nullable": true,
            "description": "Populated when status is FAILED — show it to the user verbatim."
          },
          "chunkCount": {
            "type": "number",
            "description": "Chunks embedded. Zero while pending."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "conversationId",
          "sourceType",
          "status",
          "chunkCount",
          "createdAt"
        ]
      },
      "IngestUrlDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "A YouTube link or any web page. The backend classifies it.",
            "example": "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
          },
          "conversationId": {
            "type": "string"
          }
        },
        "required": [
          "url",
          "conversationId"
        ]
      },
...
      "CreditPackDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "starter"
          },
          "credits": {
            "type": "number",
            "description": "Credits granted once the payment is captured."
          },
          "amountPaise": {
            "type": "number",
            "description": "Price in paise — integer subunits, never a float. ₹99 is 9900.",
            "example": 9900
          }
        },
        "required": [
          "id",
          "credits",
          "amountPaise"
        ]
      },
      "CreditPackId": {
        "type": "string",
        "enum": [
          "starter",
          "plus",
          "pro"
        ],
        "description": "Which pack to buy. The server owns the price."
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "packId": {
            "description": "Which pack to buy. The server owns the price.",
            "example": "starter",
            "allOf": [
              {
                "$ref": "#/components/schemas/CreditPackId"
              }
            ]
          }
        },
        "required": [
          "packId"
        ]
      },
      "OrderDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "description": "Razorpay order id. Pass as `order_id` to checkout.",
            "example": "order_MkT8xY2zAbCdEf"
          },
          "amountPaise": {
            "type": "number",
            "example": 9900
          },
          "credits": {
            "type": "number",
            "description": "Credits this purchase will grant, once captured."
          },
          "paymentId": {
            "type": "string",
            "description": "Local Payment row id. Useful for support; credits are granted by the webhook, never by the browser reporting success."
          }
        },
        "required": [
          "orderId",
          "amountPaise",
          "credits",
          "paymentId"
        ]
      }
    }
  }
}